requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
markdownify==0.14.1
Pillow==11.1.0
imagehash==4.3.2
openai==1.59.6
